appsitsolution/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.2
boto3==1.22.0
botocore==1.25.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
cssselect2==0.6.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.4
django-ckeditor==6.4.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-redis==5.2.0
django-slick-reporting==0.6.3
et-xmlfile==1.1.0
fiscalyear==0.4.0
future==0.18.2
google-auth==2.6.6
google-auth-oauthlib==0.5.1
gspread==5.3.2
html5lib==1.1
httplib2==0.20.4
idna==3.3
jmespath==1.0.0
juspayp3==1.0.1
loadtest==0.0.2
lxml==4.8.0
MarkupPy==1.14
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
oscrypto==1.3.0
packaging==21.3
paywix==1.4.0
Pillow==9.1.0
psutil==5.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyHanko==0.13.0
pyhanko-certvalidator==0.19.5
pyotp==2.6.0
pyparsing==3.0.8
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
razorpay==1.3.0
redis==4.2.2
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
wrapt==1.14.0
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
